+++ lib/kentouzu/controller.rb
@@ -17,21 +17,21 @@ module Kentouzu
     end
 
     def drafts_enabled_for_controller
-      true
+      Kentouzu.enabled?
     end
 
     private
 
     def set_drafts_source
-      ::Kentouzu.source = user_for_drafts
+      Kentouzu.source = user_for_drafts
     end
 
     def set_drafts_controller_info
-      ::Kentouzu.controller_info = info_for_drafts
+      Kentouzu.controller_info = info_for_drafts
     end
 
     def set_drafts_enabled_for_controller
-      ::Kentouzu.enabled_for_controller = drafts_enabled_for_controller
+      Kentouzu.enabled_for_controller = drafts_enabled_for_controller
     end
   end
 end