+++ lib/kentouzu/draft.rb
@@ -1,4 +1,8 @@
 class Draft < ActiveRecord::Base
+  if Kentouzu.active_record_protected_attributes?
+    attr_accessible :item_type, :item_id, :event, :source_type, :source_id, :object
+  end
+
   belongs_to :item, :polymorphic => true
 
   belongs_to :source, :polymorphic => true
@@ -6,7 +10,11 @@ class Draft < ActiveRecord::Base
   validates_presence_of :event
 
   def self.with_item_keys(item_type, item_id)
-    scoped :conditions => { :item_type => item_type, :item_id => item_id }
+    where :item_type => item_type, :item_id => item_id
+  end
+
+  def self.with_source_keys(source_type, source_id)
+    where :source_type => source_type, :source_id => source_id
   end
 
   def self.creates
@@ -17,57 +25,91 @@ class Draft < ActiveRecord::Base
     where :event => 'update'
   end
 
-  scope :subsequent, lambda { |draft| where(["#{self.primary_key} > ?", draft]).order("#{self.primary_key} ASC") }
+  def subsequent(obj, timestamp_arg = false)
+    if timestamp_arg != true && self.primary_key_is_int?
+      where(arel_table[primary_key].gt(obj.id)).order(arel_table[primary_key].asc)
+    else
+      obj = obj.send(Kentouzu.timestamp_field) if obj.is_a?(self)
 
-  scope :preceding, lambda { |draft| where(["#{self.primary_key} < ?}", draft]).order("#{self.primary_key} DESC") }
+      where(arel_table[Kentouzu.timestamp_field].gt(obj)).order(self.timestamp_sort_order)
+    end
+  end
 
-  scope :following, lambda { |timestamp| where(["#{Kentouzu.timestamp_field} > ?"], timestamp).order("#{Kentouzu.timestamp_field} ASC, #{self.primary_key} ASC") }
+  def preceding(obj, timestamp_arg = false)
+    if timestamp_arg != true && self.primary_key_is_int?
+      where(arel_table[primary_key].lt(obj.id)).order(arel_table[primary_key].asc)
+    else
+      obj = obj.send(Kentouzu.timestamp_field) if obj.is_a?(self)
 
-  scope :between, lambda { |start_time, end_time| where(["#{Kentouzu.timestamp_field} > ? AND #{Kentouzu.timestamp_field} < ?", start_time, end_time]).order("#{Kentouzu.timestamp_field} ASC, #{self.primary_key} ASC") }
+      where(arel_table[Kentouzu.timestamp_field].lt(obj)).order(self.timestamp_sort_order)
+    end
+  end
 
+  def following(timestamp)
+    where(arel_table[Kentouzu.timestamp_field].gt(timestamp)).order(self.timestamp_sort_order)
+  end
+
+  def between(start_time, end_time)
+    where(arel_tabl[Kentouzu.timestamp_field].gt(start_time).and(arel_table[Kentouzu.timestamp_field].lt(end_time))).order(self.timestamp_sort_order)
+  end
+
+  def timestamp_sort_order(direction = 'asc')
+    [arel_table[Kentouzu.timestamp_field].send(direction.downcase)].tap do |array|
+      array << arel_table[primary_key].send(direction.downcase) if self.primary_key_is_int?
+    end
+  end
+
+  # Restore the item from this draft.
+  #
+  # Options:
+  # :has_one      Set to `false` to disable has_one reification.
+  #               Set to a float to change the lookback time.
   def reify(options = {})
+    return nil if object.nil?
+
     without_identity_map do
       options[:has_one] = 3 if options[:has_one] == true
+
       options.reverse_merge! :has_one => false
 
-      unless object.nil?
-        #This appears to be necessary if for some reason the draft's model hasn't been loaded (such as when done in the console).
-        require self.item_type.underscore
+      #This appears to be necessary if for some reason the draft's model hasn't been loaded (such as when done in the console).
+      require self.item_type.underscore
 
-        loaded_object = YAML::load object
+      loaded_object = YAML::load object
 
-        if item
-          model = item
-        else
-          inheritance_column_name = item_type.constantize.inheritance_column
+      if item
+        model = item
+      else
+        inheritance_column_name = item_type.constantize.inheritance_column
 
-          class_name = loaded_object.respond_to?(inheritance_column_name.to_sym) && loaded_object.send(inheritance_column_name.to_sym).present? ? loaded_object.send(inheritance_column_name.to_sym) : item_type
+        class_name = loaded_object.respond_to?(inheritance_column_name.to_sym) && loaded_object.send(inheritance_column_name.to_sym).present? ? loaded_object.send(inheritance_column_name.to_sym) : item_type
 
-          klass = class_name.constantize
+        klass = class_name.constantize
 
-          model = klass.new
-        end
-
-        loaded_object.attributes.each do |key, value|
-          if model.respond_to?("#{key}=")
-            model.send :write_attribute, key.to_sym, value
-          else
-            logger.warn "Attribute #{key} does not exist on #{item_type} (Draft ID: #{id})."
-          end
-        end
-
-        model.send "#{model.class.draft_association_name}=", self
-
-        unless options[:has_one] == false
-          reify_has_ones model, options[:has_one]
-        end
-
-        model
+        model = klass.new
       end
+
+      loaded_object.attributes.each do |key, value|
+        if model.respond_to?("#{key}=")
+          model.send :write_attribute, key.to_sym, value
+        else
+          logger.warn "Attribute #{key} does not exist on #{item_type} (Draft ID: #{id})."
+        end
+      end
+
+      model.send "#{model.class.draft_association_name}=", self
+
+      unless options[:has_one] == false
+        reify_has_ones model, options[:has_one]
+      end
+
+      model
     end
   end
 
   def approve
+    warn 'DEPRECATED: `approve` should be handled by your application, not Kentouzu. Will be removed in Kentouzu 0.2.0.'
+
     model = self.reify
 
     if model
@@ -88,7 +130,15 @@ class Draft < ActiveRecord::Base
   end
 
   def reject
+    warn 'DEPRECATED: `reject` should be handled by your application, not Kentouzu. Will be removed in Kentouzu 0.2.0.'
+
     self.destroy
+  end
+
+  def primary_key_is_int?
+    @primary_key_is_int ||= columns_hash[primary_key].type == :integer
+  rescue
+    true
   end
 
   private